Run to user code (#762)
* Run to user code

* Fix "access violation" error

* Run to user code

* Run to user code

* Run to user code

* Run to user code

* Run to user code

* Run to user code

* Run to user code

* Run to user code

* Run to user code

* Fixed compiler warnings

* Run to user code

* Run to user code

* Run to user code

Please add an icon for it.

* Run to user code

* change the error message

* Run to user code

* Run to user code

* Update translation

* Update translation

* Enlarge the buffer size to 1KB

* Fix locking condition

* Fix race condition

* re-apply recent commits

+++ src/gui/Src/Gui/CPUSideBar.cpp
@@ -652,7 +652,7 @@ void CPUSideBar::AllocateJumpOffsets(std::vector<JumpLine> & jumpLines)
     unsigned int* numLines = new unsigned int[viewableRows];
     memset(numLines, 0, sizeof(unsigned int) * viewableRows);
     // preprocessing
-    for(int i = 0; i < jumpLines.size(); i++)
+    for(size_t i = 0; i < jumpLines.size(); i++)
     {
         JumpLine & jmp = jumpLines.at(i);
         jmp.jumpOffset = abs(jmp.destLine - jmp.line);
@@ -663,7 +663,7 @@ void CPUSideBar::AllocateJumpOffsets(std::vector<JumpLine> & jumpLines)
         return op2.jumpOffset > op1.jumpOffset;
     });
     // Allocate jump offsets
-    for(int i = 0; i < jumpLines.size(); i++)
+    for(size_t i = 0; i < jumpLines.size(); i++)
     {
         JumpLine & jmp = jumpLines.at(i);
         unsigned int maxJmpOffset = 0;
